% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVDRisk_Data.R
\docType{data}
\name{CVDRisk}
\alias{CVDRisk}
\title{CVD Risk Dataset Without Lab Data}
\format{
A data frame
}
\usage{
data(CVDRisk)
}
\description{
A dataset containing the Region, Age, Gender, Smoke, BMI, SBP, and Risk.
}
\details{
\itemize{
  \item Region. Region of residence including Andean Latin America, Australasia, Caribbean, Central Asia, Central Europe,
  Central Latin America, Central Sub-Saharan Africa, East Asia, Eastern Europe, Eastern Sub-Saharan Africa,
  High-Income Asia Pacific, High-Income North America, North Afrika and Middle East, Oceania, South Asia,
  Southeast Asia, Southern Latin America, Southern Sub-Saharan Africa, Tropical Latin America,
  Western Europe, Western Sub-Saharan Africa.
  \item Gender. Gender including "M" Male and "F" Female.
  \item Smoke. Smoke including "Smoker" and "Non-smoker"
  \item Age. Age including age groups: 40-44, 45-49, 50-54, 55-59, 60-64, 65-69, 70-74.
  \item BMI. BMI including <20, 20-24, 25-29, 30-35, >=35.
  \item SBP. SBP in (mmHg) including <120, 120-139, 140-159, 160-179, >=180.
  \item Risk. world Health Organization cardiovascular disease risk revised estimates based on 21 global regions.
}
}
\keyword{datasets}
