% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHOCVRisk.R
\name{WHOCVRisk}
\alias{WHOCVRisk}
\title{WHOCVRisk: Calculating cardiovascular disease risk based on WHO cardiovascular disease risk charts: revised models to estimate risk in 21 global regions.}
\usage{
WHOCVRisk(region, age, gender, smoke, sbp, bmi = NULL, dm = NULL, chol = NULL)
}
\arguments{
\item{region}{Region of residence including Andean Latin America, Australasia, Caribbean, Central Asia, Central Europe,
Central Latin America, Central Sub-Saharan Africa, East Asia, Eastern Europe, Eastern Sub-Saharan Africa,
High-Income Asia Pacific, High-Income North America, North Afrika and Middle East, Oceania, South Asia,
Southeast Asia, Southern Latin America, Southern Sub-Saharan Africa, Tropical Latin America,
Western Europe, Western Sub-Saharan Africa.}

\item{age}{Age from 40 to 74.}

\item{gender}{Gender including "M" Male and "F" Female.}

\item{smoke}{Smoke including "Smoker" and "Non-smoker".}

\item{sbp}{SBP in (mmHg).}

\item{bmi}{BMI (Only when lab data is not available). BMI= Weight in kilograms divided by height in meters squared.}

\item{dm}{Diabetes (Only when lab data is available) including "Yes" Diabetic and "No" Non-Diabetic.}

\item{chol}{Total cholesterol in (mmol/l) (Only when lab data is available).}
}
\value{
Output is the World Health Organization cardiovascular disease risk revised estimates based on 21 global regions.
}
\description{
Calculating cardiovascular disease risk based on WHO cardiovascular disease risk charts
}
\examples{
Example when lab data is not available

  WHOCVRisk(region="High-Income Asia Pacific",
         age=50,gender="F",smoke="Smoker",
         bmi=27,sbp=160)

Example when lab data is available

  WHOCVRisk(region="High-Income Asia Pacific",
         age=50,gender="F",smoke="Smoker",
         sbp=160,dm="Yes",chol=4.5)

}
