% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHOCVRisk_Shiny.R

\name{WHOCVRisk_Shiny}
\alias{WHOCVRisk_Shiny}
\title{WHOCVRisk_Shiny: Shiny app for Calculating cardiovascular disease risk based on WHO cardiovascular disease risk charts: revised models to estimate risk in 21 global regions.}
\usage{
WHOCVRisk_Shiny()
}

\description{
Lunch a shiny app on the local server for calculating cardiovascular disease risk
}

\examples{
WHOCVRisk_Shiny()

}
© 2022 Git
